+++ src/AppendStream.php
@@ -213,11 +213,6 @@ class AppendStream implements StreamInterface
         return false;
     }
 
-    public function flush()
-    {
-        return false;
-    }
-
     public function getMetadata($key = null)
     {
         return $key ? null : [];