+++ src/NullStream.php
@@ -71,11 +71,6 @@ class NullStream implements StreamInterface
         return strlen($string);
     }
 
-    public function flush()
-    {
-        return false;
-    }
-
     public function getMetadata($key = null)
     {
         return $key ? null : [];