+++ src/StreamInterface.php
@@ -115,13 +115,6 @@ interface StreamInterface
     public function write($string);
 
     /**
-     * Flush the write buffers of the stream.
-     *
-     * @return bool Returns true on success and false on failure
-     */
-    public function flush();
-
-    /**
      * Returns whether or not the stream is readable
      *
      * @return bool